Update README.md
-In this project, we built an Agglomerative Hierarchical Clustering model on Unsupervised data and get our dependent value. 
First, we checked our Unsupervised data and there are missing values in some columns. Before the normalization, we have to treat the missing values and drop some columns if needed. Then, we normalized our data and built a Dendogram and set a threshold distance and draw a horizontal line for selecting the number of clusters.
Now, we have dependent value, so we can clean our data as always and build a machine learning model and get predictions.